auth-service/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
scipy>=1.10.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
